Added sub-command functions.
Now, the parser sets a function to be called, so that the main function is
essentially subtype-polymorphic. Now, we need to get crypto working for
ID hashes, and yaml for serialzation formats.

Basically, tag lists and file lists will be stored in YAML lists, in
sorted order. That way, we can use merges to assemble combined lists
quickly. We determine the name of the list file by hashing its contents.
This means that a file is hashed to produce the file that holds its tags,
and a tag is hashed to produce the file that holds its list of paths.

We will store all tag and path files in a .tags subdir under CWD, or if
TAGDIR is set, then there. This will allow the user to maintain several
tag DBs, in different locations, and switch by either cd'ing or setting the
environment variable.