Swapped get and tags.
Oops! I put tags in get! Anyway, swapped them out, and simplified tags,
in light of the fact that we are only taking one file as argument.

As for get, we reduce union over the sets, and fromMaybe ensures that at
worst, we get empty sets. Now, we must do two things: first, we must fix
set so that, in addition to registering the file with tags, we register
the tags with the file. Second, we must add expression processing to the
get function.