Stripped Monad earlier.
While the monad is a good thing, it is not a good thing to confuse types. The
monad returns Nothing on either failure to open a file or missing key. This
is fine, but we actually want an empty set in these cases. Therefore, we need
to make sure to convert all nothing results to set().